# docs/swagger.yaml
definitions:
...
  request.TenantReq:
    properties:
      accessExpire:
        type: integer
      codeExpire:
        type: integer
      company:
        type: string
      describe:
        type: string
      grantTypes:
        items:
          type: string
        type: array
      host:
        type: string
      idExpire:
        type: integer
      isCode:
        type: boolean
      isCredential:
        type: boolean
      isDeviceFlow:
        type: boolean
      isDisabled:
        type: boolean
      isPassword:
        type: boolean
      isRefresh:
        type: boolean
      name:
        type: string
      refreshExpire:
        type: integer
      type:
        type: integer
      userPoolId:
        type: integer
    required:
    - company
    - host
    - name
    - userPoolId
    type: object
  request.UserPoolReq:
    properties:
      describe:
        type: string
      name:
        type: string
    type: object
  request.UserReq:
    properties:
      displayName:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      type:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
  license:
    name: MIT
    url: https://github.com/hello-jiangxiaoyu/QuickAuth/blob/main/LICENSE
  title: Quick Auth 接口文档
  version: "1.0"
paths:
...
  /api/quick/apps/{appId}/tenants:
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: get tenant list
      tags:
      - tenant
    post:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.TenantReq'
      responses:
        "200":
          description: OK
      summary: create tenant
      tags:
      - tenant
  /api/quick/apps/{appId}/tenants/{tenantId}:
    delete:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      - description: tenant id
        in: path
        name: tenantId
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: delete tenant
      tags:
      - tenant
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      - description: tenant id
        in: path
        name: tenantId
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: get tenant details
      tags:
      - tenant
    put:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      - description: tenant id
        in: path
        name: tenantId
        required: true
        type: integer
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.TenantReq'
      responses:
        "200":
          description: OK
      summary: modify tenant
      tags:
      - tenant
...
  /api/quick/user-pools:
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      responses:
        "200":
          description: OK
      summary: list user pool
      tags:
      - user
    post:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.UserPoolReq'
      responses:
        "200":
          description: OK
      summary: create user pool
      tags:
      - user
  /api/quick/user-pools/{poolId}:
    delete:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: user pool id
        in: path
        name: poolId
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: delete user pool
      tags:
      - user
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: user pool id
        in: path
        name: poolId
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: get user pool info
      tags:
      - user
    put:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: user pool id
        in: path
        name: poolId
        required: true
        type: integer
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.UserPoolReq'
      responses:
        "200":
          description: OK
      summary: modify user pool
      tags:
      - user
  /api/quick/user-pools/{poolId}/users:
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: user pool id
        in: path
        name: poolId
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: list users
      tags:
      - user
    post:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: user pool id
        in: path
        name: poolId
        required: true
        type: integer
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.UserReq'
      responses:
        "200":
          description: OK
      summary: create user
      tags:
      - user
  /api/quick/user-pools/{poolId}/users/{userId}:
    delete:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: user pool id
        in: path
        name: poolId
        required: true
        type: integer
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete user
      tags:
      - user
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: user pool id
        in: path
        name: poolId
        required: true
        type: integer
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: get user info
      tags:
      - user
    put:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: user pool id
        in: path
        name: poolId
        required: true
        type: integer
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.UserReq'
      responses:
        "200":
          description: OK
      summary: modify user
      tags:
      - user
securityDefinitions:
  Login:
    in: header
    name: token
    type: apiKey
swagger: "2.0"